Refactor Action Manager initialization for the Editor Main Window into a helper class.
