+++ Code/Editor/MainWindow.h
@@ -28,6 +28,8 @@
 #include <AzToolsFramework/SourceControl/SourceControlAPI.h>
 
 #include "IEditor.h"
+
+#include <Core/EditorActionsHandler.h>
 #endif
 
 class ActionManager;
@@ -59,9 +61,6 @@ namespace AzQtComponents
 
 namespace AzToolsFramework
 {
-    class ActionManagerInterface;
-    class MenuManagerInterface;
-    class ToolBarManagerInterface;
     class Ticker;
     class QtSourceControlNotificationHandler;
 
@@ -213,16 +212,6 @@ private:
 
     QToolButton* CreateUndoRedoButton(int command);
 
-    // Editor Action Manager initialization functions
-    AzToolsFramework::ActionManagerInterface* m_actionManagerInterface = nullptr;
-    AzToolsFramework::MenuManagerInterface* m_menuManagerInterface = nullptr;
-    AzToolsFramework::ToolBarManagerInterface* m_toolBarManagerInterface = nullptr;
-
-    void InitializeActionContext();
-    void InitializeActions();
-    void InitializeMenus();
-    void InitializeToolBars();
-
 private Q_SLOTS:
     void ShowKeyboardCustomization();
     void ExportKeyboardShortcuts();
@@ -277,6 +266,7 @@ private:
     static MainWindow* m_instance;
 
     AzQtComponents::DockMainWindow* m_viewPaneHost;
+    EditorActionsHandler m_editorActionsHandler;
 
     QTimer* m_autoSaveTimer;
     QTimer* m_autoRemindTimer;
